// frontend/src/util/payload-rest-client/types.ts
export type ResetPasswordResult<T> = {
    message: string;
    user: T;
    token: string;
    errors?: string[];
};